Move config code to new _shared package
