+++ src/dapla_metadata/_shared/__init__.py
@@ -1,0 +1,1 @@
+"""Utility code intended to be used by multiple packages within this project, but not by end users."""
