+++ src/dapla_metadata/datasets/utility/enums.py
@@ -5,25 +5,6 @@ from __future__ import annotations
 from enum import Enum
 
 
-class DaplaRegion(str, Enum):
-    """Dapla platforms/regions."""
-
-    DAPLA_LAB = "DAPLA_LAB"
-    BIP = "BIP"
-    ON_PREM = "ON_PREM"
-    CLOUD_RUN = "CLOUD_RUN"
-
-
-class DaplaService(str, Enum):
-    """Dapla services."""
-
-    DATADOC = "DATADOC"
-    JUPYTERLAB = "JUPYTERLAB"
-    VS_CODE = "VS_CODE"
-    R_STUDIO = "R_STUDIO"
-    KILDOMATEN = "KILDOMATEN"
-
-
 class SupportedLanguages(str, Enum):
     """The list of languages metadata may be recorded in.
