+++ tests/datasets/constants.py
@@ -48,14 +48,6 @@ TEST_PROCESSED_DATA_POPULATION_DIRECTORY = TEST_RESOURCES_DIRECTORY / "klargjort
 
 CODE_LIST_DIR = "code_list"
 
-JUPYTERHUB_USER = "JUPYTERHUB_USER"
-
-DAPLA_REGION = "DAPLA_REGION"
-
-DAPLA_GROUP_CONTEXT = "DAPLA_GROUP_CONTEXT"
-
-DAPLA_SERVICE = "DAPLA_SERVICE"
-
 DATADOC_METADATA_MODULE = "dapla_metadata.datasets"
 
 DATADOC_METADATA_MODULE_CORE = "dapla_metadata.datasets.core"
