+++ tests/datasets/test_datadoc_metadata.py
@@ -25,12 +25,12 @@ from datadoc_model.model import IsPersonalData
 from datadoc_model.model import Variable
 from datadoc_model.model import VariableRole
 
+from dapla_metadata._shared.user_info import PLACEHOLDER_EMAIL_ADDRESS
+from dapla_metadata._shared.user_info import TestUserInfo
 from dapla_metadata.datasets.core import Datadoc
 from dapla_metadata.datasets.core import InconsistentDatasetsError
 from dapla_metadata.datasets.core import InconsistentDatasetsWarning
 from dapla_metadata.datasets.statistic_subject_mapping import StatisticSubjectMapping
-from dapla_metadata.datasets.user_info import PLACEHOLDER_EMAIL_ADDRESS
-from dapla_metadata.datasets.user_info import TestUserInfo
 from dapla_metadata.datasets.utility.constants import (
     DATASET_FIELDS_FROM_EXISTING_METADATA,
 )
