+++ tests/utils/constants.py
@@ -1,3 +1,7 @@
 VARDEF_EXAMPLE_DEFINITION_ID = "wypvb3wd"
 VARDEF_EXAMPLE_ACTIVE_GROUP = "dapla-felles-developers"
 VARDEF_EXAMPLE_DATE = "1970-01-01"
+JUPYTERHUB_USER = "JUPYTERHUB_USER"
+DAPLA_REGION = "DAPLA_REGION"
+DAPLA_GROUP_CONTEXT = "DAPLA_GROUP_CONTEXT"
+DAPLA_SERVICE = "DAPLA_SERVICE"
